rename to pool, more tests
