+++ rayoff/src/lib.rs
@@ -1,2 +1,2 @@
 mod job;
-pub mod rayoff;
+pub mod pool;
